Final modifications for PR #42
