+++ mdss/__init__.py
@@ -1,13 +1,12 @@
 __version__ = "0.1.0"
 
 from .src import main
-from .utils import utils
-from .src.main import simulation
-from .src.main import post_process
+from .utils import tools
+from .src.main import simulation, custom_sim, post_process
 from .src.main_helper import execute
 try:
     from .src.aerostruct import Problem # Requires mphys and other libraries. Causes problems when using in a different environment.
 except:
     pass
 
-__all__ = ["main", "utils", "Problem", "simulation", "execute", "post_process"]
+__all__ = ["main", "tools", "Problem", "simulation", "custom_sim", "execute", "post_process"]
