+++ mdss/resources/aerostruct_defaults.py
@@ -11,15 +11,11 @@ default_aero_options_aerostructural = {
     "outputDirectory": '.', 
     "monitorvariables": ["resrho", "resturb", "cl", "cd", "yplus"],
     "writeTecplotSurfaceSolution": True,
+    "solutionPrecision": "double", #  Best for restart
+    "volumeVariables": ['resrho', 'mach'],
     # Physics Parameters
     "equationType": "RANS",
     "liftindex": 3,  # z is the lift direction
-    # Solver Parameters
-    "smoother": "DADI",
-    "CFL": 1.5,
-    "CFLCoarse": 1.25,
-    "MGCycle": "sg",
-    "MGStartLevel": -1,
     # ANK Solver Parameters
     "useANKSolver": True,
     # Termination Criteria
