+++ mdss/src/__init__.py
@@ -1,8 +1,8 @@
 # src/__init__.py
-from .main import simulation, post_process
+from .main import simulation, custom_sim, post_process
 try:
     from .aerostruct import Problem
 except:
     pass
 
-__all__ = ["simulation", "post_process", "Problem"]
+__all__ = ["simulation", "custom_sim", "post_process", "Problem"]
