+++ mdss/utils/tools.py
@@ -1,16 +1,7 @@
-import copy
-
-from mpi4py import MPI
 from enum import Enum
-import os
-import shutil
-import yaml
-import importlib.resources as pkg_resources
-import random
-from pydantic import BaseModel
-from typing import Optional, Literal
+import os, yaml
+from mpi4py import MPI
 
-from mdss.src.main import simulation
 from mdss.utils.helpers import *
 from mdss.resources.yaml_config import check_input_yaml, ref_scenario_info
 
@@ -231,141 +222,4 @@ def get_sim_data(yaml_input):
                     sim_data[hierarchy_name][case_name][scenario_name][mesh_file]['failed_aoa'] = failed_aoa
 
     return sim_data
-
-################################################################################
-# Functions to run predetermined simulations
-################################################################################
-
-class ref_case_info(BaseModel):
-    out_dir: str=None
-    meshes_folder_path: Optional[str]=None
-    mesh_files: Optional[list[str]]=None
-    aoa_list: list[float]
-    aero_options: Optional[dict]=None
-    struct_options: Optional[dict]=None
-    
-class custom_sim(simulation):
-    """
-    Class to Run a predefined case simulation
-
-    It sets up and executes a simulation based on the provided case information.
-    It validates the input, prepares a temporary directory for files, and cleans up after the run.
-
-    Parameters
-    ----------
-    yaml_input: str
-        The YAML file path or raw YAML string for the simulation.
-
-    Notes
-    ------
-    - Creates a temporary directory for the simulation input and output files.
-    - Deletes the temporary directory after the simulation run.
-    """
-    def __init__(self, yaml_input:str, out_dir:str=None):
-        self.yaml_input = yaml_input
-        super().__init__(yaml_input)  # Leverages validation, parsing, and setup from `simulation`
-        if comm.rank == 0:
-            if os.path.exists(self.sim_info['out_dir']) and not os.listdir(self.sim_info['out_dir']):  # Check if the output directory exits and is empty
-                os.rmdir(self.sim_info['out_dir'])  # Remove the directory only if it is empty
-    
-    def run(self, case_info):
-        """
-        Parameters
-        ----------
-        case_info: dict
-            A dictionary containing the case information. It should follow the structure defined by the `ref_case_info` class:
-            
-            - `out_dir` (str): Path to the output directory.
-            - `meshes_folder_path` (str): Path to the directory containing mesh files.
-            - `mesh_files` (list[str]): List of mesh file names.
-            - `aoa_list` (list[float]): List of angles of attack for the simulation.
-            - `aero_options` (Optional[dict]): Dictionary containing ADflow solver parameters (optional).
-            - `struct_options` (Optional[dict]): Dictionary containing structural info for aero structural problem (optional).
-        """
-        ref_case_info.model_validate(case_info)
-        
-        case = self.sim_info['hierarchies'][0]['cases'][0]
-        case_info['aoa_list'] = [float(aoa) for aoa in case_info['aoa_list']]  # Ensures all angles of attack to float and converts a numpy array to a list
-        # Extract only fields that were explicitly provided (non-None)
-        for key, value in case_info.items():
-            if value is None or key in ['out_dir', 'aoa_list']:
-                continue  # Skip unset or None fields
-            elif key in {'aero_options', 'struct_options'}:
-                case.setdefault(key, {}).update(value)
-            else:
-                case[key] = value
-
-        # Update the angle of attack in the scenario and scenario information
-        scenario = case['scenarios'][0]
-        scenario['aoa_list'] = case_info['aoa_list']
-        ref_scenario_info.model_validate(scenario)
-        # Update the scenario information in the case
-        for key, value in scenario.items():
-            if value is None or key in ['name', 'aoa_list', 'exp_data']:
-                continue
-            else:
-                scenario[key] = float(value)  # Convert all the other values to float
-        
-        
-        if comm.rank == 0:
-            randn = random.randint(1000, 9999)
-        else:
-            # Other processes initialize the variable
-            randn = None
-        
-        # Broadcast the random number to all processes
-        randn = comm.bcast(randn, root=0)
-        cwd = os.getcwd()
-        if 'out_dir' in case_info.keys():
-            if not os.path.exists(case_info['out_dir']) and comm.rank == 0:
-                os.mkdir(case_info['out_dir'])
-            self.sim_info['out_dir'] = case_info['out_dir']
-            self.out_dir = case_info['out_dir']
-        else:
-            temp_dir = os.path.join(cwd, f"temp_{randn}")
-            msg = f"Creating a temporary folder to run simulations: {temp_dir}"
-            print_msg(msg, 'notice', comm)
-            if comm.rank == 0:
-                os.mkdir(temp_dir)
-            self.sim_info['out_dir'] = temp_dir
-            self.out_dir = temp_dir
-        comm.Barrier()
-        self.final_out_file = os.path.join(self.out_dir, "overall_sim_info.yaml")  # Set the final output file path
-        modified_yaml_input = yaml.dump(self.sim_info, sort_keys=False)
-        check_input_yaml(modified_yaml_input)  # Validate the modified YAML input
-        # Write the modified YAML input to a file
-        if not os.path.exists(self.out_dir):
-            if comm.rank == 0:
-                os.mkdir(self.out_dir)
-        self.info_file = os.path.join(self.out_dir, "input.yaml")  # Set the path for the input YAML file
-        with open(self.info_file, 'w') as f:
-            yaml.dump(self.sim_info, f, sort_keys=False)
-
-        if self.machine_type == MachineType.HPC:
-            self.wait_for_job = True # To toggle to wait for the job to finish.
-            
-        # Call the parent class's run method to execute the simulation
-        super().run()  
-
-        comm.Barrier() 
-
-        # Read the simulation data from the final output file
-        if os.path.exists(self.final_out_file):
-            sim_data = get_sim_data(self.final_out_file)
-        else:
-            raise FileNotFoundError(f"Expected output file not found: {self.final_out_file}")
-
-        sim_data = comm.bcast(sim_data, root=0)
-        comm.Barrier()  # Ensure all ranks are done reading
-        
-        # Cleanup temp directory if created
-        if 'out_dir' not in case_info and comm.rank == 0:
-            shutil.rmtree(self.out_dir)
-
-        # Reset sim_info
-        self.sim_info, self.yaml_input_type = load_yaml_input(self.yaml_input, comm)
-        self.sim_info['out_dir'] = os.path.abspath(self.sim_info['out_dir'])
-        self.out_dir = self.sim_info['out_dir']
-        
-        return sim_data
